+++ main/src/vtr_lidar/src/modules/planning/change_detection_module_v3.cpp
@@ -531,8 +531,7 @@ void ChangeDetectionModuleV3::run_(QueryCache &qdata0, OutputCache &output0,
     //CLOG(ERROR, "obstacle_detection.cbit") << "Displaying all Values: " << vals2;
 
     // Update the output cache
-    output.costmap_sid = dense_costmap->vertex_sid();
-    CLOG(ERROR, "obstacle_detection.cbit") << "costmap_sid" << dense_costmap->vertex_sid(); 
+    output.costmap_sid = dense_costmap->vertex_sid(); 
     output.obs_map =  dense_costmap->filter(0.01); //was config_->costmap_filter_value
     output.grid_resolution = config_->resolution;
