+++ main/src/vtr_path_planning/include/vtr_path_planning/path_planning.hpp
@@ -17,7 +17,6 @@
  * \author Yuchen Wu, Autonomous Space Robotics Lab (ASRL)
  */
 #pragma once
-#include "vtr_path_planning/mpc/mpc_path_planner2.hpp"
-//#include "vtr_path_planning/mpc/mpc_path_planner.hpp"
+#include "vtr_path_planning/mpc/mpc_path_planner.hpp"
 #include "vtr_path_planning/cbit/cbit.hpp"
 #include "vtr_path_planning/stationary_planner.hpp"
