+++ main/src/vtr_path_planning/src/mpc/mpc_path_planner.cpp
@@ -13,20 +13,61 @@
 // limitations under the License.
 
 /**
- * \file mpc.cpp
+ * \file mpc_path_planner.cpp
  * \author Jordy Sehn, Autonomous Space Robotics Lab (ASRL)
  */
+
 #include "vtr_path_planning/mpc/mpc_path_planner.hpp"
-//#include "vtr_lidar/cache.hpp"
+#include "vtr_path_planning/mpc/lateral_error_evaluators.hpp"
+#include "vtr_path_planning/mpc/custom_loss_functions.hpp"
+#include "vtr_path_planning/mpc/scalar_log_barrier_evaluator.hpp"
 
-// This file is used to generate a tracking mpc output velocity command given a discretized path to follow and optimization parameters
-// It is used in cbit.cpp in both the vtr_lidar package (obstacle avoidance) and vtr_path_planning packages (obstacle free) in the computeCommand function
 
 
-// Main MPC problem solve function
-struct mpc_result SolveMPC(Eigen::Matrix<double, 2, 1> previous_vel, lgmath::se3::Transformation T0, std::vector<lgmath::se3::Transformation> measurements, int K, double DT, double VF, Eigen::Matrix<double, 6, 6> pose_noise_vect, Eigen::Matrix<double, 2, 2> vel_noise_vect, Eigen::Matrix<double, 2, 2> accel_noise_vect, Eigen::Matrix<double, 6, 6> kin_noise_vect, bool point_stabilization)
+// Main MPC problem solve function - TODO: dump all these arguments into an mpc config class
+struct mpc_result SolveMPC(Eigen::Matrix<double, 2, 1> previous_vel, lgmath::se3::Transformation T0, std::vector<lgmath::se3::Transformation> measurements, std::vector<lgmath::se3::Transformation> measurements_cbit, std::vector<double> barrier_q_left, std::vector<double> barrier_q_right, int K, double DT, double VF, Eigen::Matrix<double, 1, 1> lat_noise_vect, Eigen::Matrix<double, 6, 6> pose_noise_vect, Eigen::Matrix<double, 2, 2> vel_noise_vect, Eigen::Matrix<double, 2, 2> accel_noise_vect, Eigen::Matrix<double, 6, 6> kin_noise_vect, bool point_stabilization, double pose_error_weight, double vel_error_weight, double acc_error_weight, double kin_error_weight, double lat_error_weight)
 {
     
+    // Pre-process the measurements to correct the cbit path yaw values depending on if the robot should be moving in forward or reverse
+    //Eigen::Matrix<double, 4, 4> T_ROT_YAW;
+    //T_ROT_YAW << -1, 0, 0, 0,
+    //            0, -1, 0, 0,
+    //            0, 0, 1, 0,
+    //            0, 0, 0, 1;
+    //lgmath::se3::Transformation T_ROT_YAW2(T_ROT_YAW);
+    //CLOG(ERROR, "mpc_debug.cbit") << "THE PI ROTATION MATRIX IS: " << T_ROT_YAW2;
+
+    //for (int i = 0; i < measurements_cbit.size(); i++)
+    //{
+    //  auto test1 = measurements[i].matrix().block(0, 0, 3, 3);
+    //  auto test2 = measurements_cbit[i].matrix().block<3, 1>(0, 3);
+    //  CLOG(ERROR, "mpc_debug.cbit") << "TEST2 IS: " << test2;
+    //  auto cbit_pose = measurements_cbit[i].matrix().block(0, 0, 3, 3);
+    //  auto teach_pose = measurements[i].matrix().block(0, 0, 3, 3);
+    //  CLOG(ERROR, "mpc_debug.cbit") << "THE CBIT MATRIX IS: " << measurements_cbit[i];
+    //  CLOG(ERROR, "mpc_debug.cbit") << "THE TEACH MATRIX IS: " << measurements[i];
+      //measurements_cbit[i] = measurements[i];
+    //  lgmath::se3::Transformation new_meas_cbit(test1,test2);
+    //  measurements_cbit[i] = new_meas_cbit;
+
+    //  CLOG(ERROR, "mpc_debug.cbit") << "THE ROTATED CBIT MATRIX IS: " << measurements_cbit[i];
+      //auto rel_pose = teach_pose.transpose() * cbit_pose;
+      //CLOG(ERROR, "mpc_debug.cbit") << "THE CBIT ROTATION MATRIX IS: " << cbit_pose;
+      //CLOG(ERROR, "mpc_debug.cbit") << "THE TEACH ROTATION MATRIX IS: " << teach_pose;
+      //CLOG(ERROR, "mpc_debug.cbit") << "THE RELATIVE ROTATION MATRIX IS: " << teach_pose.transpose() * cbit_pose;
+      //double relative_yaw = atan2(rel_pose(1,0),rel_pose(0,0));
+      //CLOG(ERROR, "mpc_debug.cbit") << "THE RELATIVE YAW IS: " << relative_yaw;
+      //if (fabs(relative_yaw) >= 1.57075)
+      //{
+      //  measurements_cbit[i] = measurements_cbit[i] * T_ROT_YAW2;
+      //  CLOG(ERROR, "mpc_debug.cbit") << "CORRECTED THE YAW";
+      //}
+      //CLOG(ERROR, "mpc_debug.cbit") << "THE ROTATED CBIT MATRIX IS: " << measurements_cbit[i];
+    //}
+    // End of pre-pre-processing measurement rotations
+
+
+
     // Conduct an MPC Iteration given current configurations
 
     // Velocity set-points (desired forward velocity and angular velocity), here we set a static forward target velocity, and try to minimize rotations (0rad/sec)
@@ -46,35 +87,28 @@ struct mpc_result SolveMPC(Eigen::Matrix<double, 2, 1> previous_vel, lgmath::se3
                 0, -1;
 
     // Lateral constraint projection matrices (Experimental)
-    Eigen::Matrix<double, 4, 1> I_4;
+    Eigen::Matrix<double, 3, 1> I_4; // In steam, the homopoint vars automatically add the 4th row 1, so representing I_4 just needs the 3 zeros
     I_4 << 0,
            0,
-           0,
-           1;
-    Eigen::Matrix<double, 1, 4> I_2_tran;
-    I_2_tran << 0, 1, 0, 0;
+           0;
 
 
     // Setup shared loss functions and noise models for all cost terms
-    const auto sharedLossFunc = steam::L2LossFunc::MakeShared();
+    //const auto sharedLossFunc = steam::L2LossFunc::MakeShared(); // The default L2 loss function weights all cost terms with a value of 1.0 (not using this one anymore)
 
-    const auto sharedPoseNoiseModel =
-        steam::StaticNoiseModel<6>::MakeShared(pose_noise_vect);
+    // The custom L2WeightedLossFunc allows you to dynamically set the weights of cost terms by providing the value as an argument
+    const auto poseLossFunc = steam::L2WeightedLossFunc::MakeShared(pose_error_weight);
+    const auto velLossFunc = steam::L2WeightedLossFunc::MakeShared(vel_error_weight);
+    const auto accelLossFunc = steam::L2WeightedLossFunc::MakeShared(acc_error_weight);
+    const auto kinLossFunc = steam::L2WeightedLossFunc::MakeShared(kin_error_weight);
+    const auto latLossFunc = steam::L2WeightedLossFunc::MakeShared(lat_error_weight); 
 
-    const auto sharedVelNoiseModel =
-        steam::StaticNoiseModel<2>::MakeShared(vel_noise_vect);
-
-    const auto sharedAccelNoiseModel =
-        steam::StaticNoiseModel<2>::MakeShared(accel_noise_vect);
-
-    const auto sharedKinNoiseModel =
-        steam::StaticNoiseModel<6>::MakeShared(kin_noise_vect);
-
-    // Experimental lat constraint (hardcoded for now)
-    Eigen::Matrix<double, 1, 1> lat_noise_vect;
-    lat_noise_vect << 1.0;
-    const auto sharedLatNoiseModel =
-        steam::StaticNoiseModel<1>::MakeShared(lat_noise_vect);
+    // Cost term Noise Covariance Initialization
+    const auto sharedPoseNoiseModel = steam::StaticNoiseModel<6>::MakeShared(pose_noise_vect);
+    const auto sharedVelNoiseModel = steam::StaticNoiseModel<2>::MakeShared(vel_noise_vect);
+    const auto sharedAccelNoiseModel = steam::StaticNoiseModel<2>::MakeShared(accel_noise_vect);
+    const auto sharedKinNoiseModel = steam::StaticNoiseModel<6>::MakeShared(kin_noise_vect);
+    const auto sharedLatNoiseModel = steam::StaticNoiseModel<1>::MakeShared(lat_noise_vect);
 
 
 
@@ -86,123 +120,229 @@ struct mpc_result SolveMPC(Eigen::Matrix<double, 2, 1> previous_vel, lgmath::se3
     lgmath::se3::Transformation T0_inv = T0.inverse();
     Eigen::Vector2d v0(0.0, 0.0);
 
-    // Pushback the initial states (current robot state)
-    pose_states.push_back(T0_inv); // Change this to T0 when implementing on robot, T0_1 for debug
-    //vel_states.push_back(std::vector<double> {0.0, 0.0}); //I think a single line way t odo this is something like Eigen::Matrix<double, 2, 1>::Zero()
+    // Push back the initial states (current robot state)
+    pose_states.push_back(T0_inv);
     vel_states.push_back(v0);
 
-    // Set the remaining states
-    for (int i=0; i<K-1; i++)
+    //CLOG(DEBUG, "mpc_debug.cbit") << "Verifying that the initial robot state is about the same as the initial reference measurement";
+    //CLOG(DEBUG, "mpc_debug.cbit") << "T0_inv: " << T0_inv;
+    //CLOG(ERROR, "mpc_debug.cbit") << "1st meas: " << measurements_cbit[0];
+
+    // Set the remaining states using a warm start from the cbit solution
+    for (int i=1; i<K; i++)
     {
-        //pose_states.push_back(lgmath::se3::Transformation()); // I wonder if here I should initialize all the states as the initial state
-        pose_states.push_back(T0_inv);
+        //pose_states.push_back(lgmath::se3::Transformation());
+        //pose_states.push_back(T0_inv); // old initialization - set all initial states to that of the current robot state
+        pose_states.push_back(measurements_cbit[i]); // New initialization - use the reference measurements from the cbit solution as our initialization - the first one is the same as our initial state
         vel_states.push_back(v0);
     }
 
     // Create Steam states
     std::vector<steam::se3::SE3StateVar::Ptr> pose_state_vars;
     std::vector<steam::vspace::VSpaceStateVar<2>::Ptr> vel_state_vars;
+    std::vector<steam::se3::SE3StateVar::Ptr> measurement_vars; // This one is for storing measurements as locked evaluators for barrier constraints
+    
+    // Create a locked state var for the 4th column of the identity matrix (used in state constraint)
+    steam::stereo::HomoPointStateVar::Ptr I_4_eval = steam::stereo::HomoPointStateVar::MakeShared(I_4);
+    I_4_eval->locked() = true;
 
+    // Create Steam variables
     for (int i = 0; i < K; i++)
     {
-        pose_state_vars.push_back(steam::se3::SE3StateVar::MakeShared(pose_states[i]));
+        pose_state_vars.push_back(steam::se3::SE3StateVar::MakeShared(pose_states[i])); 
         vel_state_vars.push_back(steam::vspace::VSpaceStateVar<2>::MakeShared(vel_states[i])); 
     }
 
-    // Lock the first (current robot) state from being able to be modified during the optimization
+    // Lock the first (current robot) state from being modified during the optimization
     pose_state_vars[0]->locked() = true;
+
 
 
     // Setup the optimization problem
     steam::OptimizationProblem opt_problem;
-    for (int i=0; i<K; i++)
+    for (int i=1; i<K; i++) // start at 1 so as to not add the locked state variable
     {
         opt_problem.addStateVariable(pose_state_vars[i]);
-        //opt_problem.addStateVariable(vel_state_vars[i]);
     }
 
+    // The velocity states should have one less variable then the pose states
     for (int i=0; i<K-1; i++)
     {
-        //opt_problem.addStateVariable(pose_state_vars[i]);
         opt_problem.addStateVariable(vel_state_vars[i]);
     }
 
     // Generate the cost terms using combinations of the built-in steam evaluators
+    double dynamic_pose_error_weight = pose_error_weight;
     for (int i = 0; i < K; i++)
     {
-        // Pose Error
+      // Pose Error
+      if (i > 0)
+      {
         const auto pose_error_func = steam::se3::SE3ErrorEvaluator::MakeShared(pose_state_vars[i], measurements[i]);
-        const auto pose_cost_term = steam::WeightedLeastSqCostTerm<6>::MakeShared(pose_error_func, sharedPoseNoiseModel, sharedLossFunc);
+        auto dynamicposeLossFunc = steam::L2WeightedLossFunc::MakeShared(dynamic_pose_error_weight);
+        const auto pose_cost_term = steam::WeightedLeastSqCostTerm<6>::MakeShared(pose_error_func, sharedPoseNoiseModel, dynamicposeLossFunc);
         opt_problem.addCostTerm(pose_cost_term);
+        dynamic_pose_error_weight = dynamic_pose_error_weight * 0.95;
+      }
 
-        // Non-Zero Velocity Penalty (OLD, not using this way any more, though might change to this when approaching end of path)
-        //const auto vel_cost_term = steam::WeightedLeastSqCostTerm<2>::MakeShared(vel_state_vars[i], sharedVelNoiseModel, sharedLossFunc);
-        //opt_problem.addCostTerm(vel_cost_term);
-
-        
-
-        // Kinematic constraints (softened but penalized heavily)
-        if (i < (K-1))
-        {
+      // Kinematic constraints (softened but penalized heavily)
+      if (i < (K-1))
+      {
         const auto lhs = steam::se3::ComposeInverseEvaluator::MakeShared(pose_state_vars[i+1], pose_state_vars[i]);
         const auto vel_proj = steam::vspace::MatrixMultEvaluator<6,2>::MakeShared(vel_state_vars[i], P_tran); // TODO, I guess this version of steam doesnt have this one, will need to do it myself
         const auto scaled_vel_proj = steam::vspace::ScalarMultEvaluator<6>::MakeShared(vel_proj, DT);
         const auto rhs = steam::se3::ExpMapEvaluator::MakeShared(scaled_vel_proj);
         const auto kin_error_func = steam::se3::LogMapEvaluator::MakeShared(steam::se3::ComposeInverseEvaluator::MakeShared(lhs, rhs));
-        const auto kin_cost_term = steam::WeightedLeastSqCostTerm<6>::MakeShared(kin_error_func, sharedKinNoiseModel, sharedLossFunc);
+        const auto kin_cost_term = steam::WeightedLeastSqCostTerm<6>::MakeShared(kin_error_func, sharedKinNoiseModel, kinLossFunc);
         opt_problem.addCostTerm(kin_cost_term);
 
+        // Non-Zero Velocity Penalty (penalty of non resting control effort helps with point stabilization)
+        //const auto vel_cost_term = steam::WeightedLeastSqCostTerm<2>::MakeShared(vel_state_vars[i], sharedVelNoiseModel, velLossFunc);
+        //opt_problem.addCostTerm(vel_cost_term);
 
-        // Experimental velocity set-point constraint (instead of non zero velocity penalty)
+        // Velocity set-point constraint - No longer using this due to complications when repeating a path in reverse
         // Only add this cost term if we are not in point stabilization mode (end of path)
+        
         //if (point_stabilization == false)
         //{
-        //  const auto vel_cost_term = steam::WeightedLeastSqCostTerm<2>::MakeShared(steam::vspace::VSpaceErrorEvaluator<2>::MakeShared(vel_state_vars[i],v_ref), sharedVelNoiseModel, sharedLossFunc);
+        //  const auto vel_cost_term = steam::WeightedLeastSqCostTerm<2>::MakeShared(steam::vspace::VSpaceErrorEvaluator<2>::MakeShared(vel_state_vars[i],v_ref), sharedVelNoiseModel, velLossFunc);
         //  opt_problem.addCostTerm(vel_cost_term);
         //}
 
-
-        // Experimental acceleration limits
+        // Experimental End of Path Termination Constraint
+        if (point_stabilization == true)
+        {
+          const auto vel_cost_term = steam::WeightedLeastSqCostTerm<2>::MakeShared(vel_state_vars[i], sharedVelNoiseModel, velLossFunc);
+          opt_problem.addCostTerm(vel_cost_term);
+        }
         
+        
+        // Acceleration Constraints
         if (i == 0)
         {
         // On the first iteration, we need to use an error with the previously applied control command state
-        const auto accel_cost_term = steam::WeightedLeastSqCostTerm<2>::MakeShared(steam::vspace::VSpaceErrorEvaluator<2>::MakeShared(vel_state_vars[i], previous_vel), sharedAccelNoiseModel, sharedLossFunc);
+        const auto accel_cost_term = steam::WeightedLeastSqCostTerm<2>::MakeShared(steam::vspace::VSpaceErrorEvaluator<2>::MakeShared(vel_state_vars[i], previous_vel), sharedAccelNoiseModel, accelLossFunc);
         opt_problem.addCostTerm(accel_cost_term);
         } 
         else
         {
         // Subsequent iterations we make an error between consecutive velocities. We penalize large changes in velocity between time steps
         const auto accel_diff = steam::vspace::AdditionEvaluator<2>::MakeShared(vel_state_vars[i], steam::vspace::NegationEvaluator<2>::MakeShared(vel_state_vars[i-1]));
-        const auto accel_cost_term = steam::WeightedLeastSqCostTerm<2>::MakeShared(accel_diff, sharedAccelNoiseModel, sharedLossFunc);
+        const auto accel_cost_term = steam::WeightedLeastSqCostTerm<2>::MakeShared(accel_diff, sharedAccelNoiseModel, accelLossFunc);
         opt_problem.addCostTerm(accel_cost_term);
-        }   
+        }  
         
-        }
+        
+      }
 
+      // Laterial Barrier State Constraints
+      if (i >= 0)
+      {
+        // Generate a locked transform evaluator to store the current measurement for state constraints
+        // The reason we make it a variable and lock it is so we can use the built in steam evaluators which require evaluable inputs
+        measurement_vars.push_back(steam::se3::SE3StateVar::MakeShared(measurements[i]));
+        measurement_vars[i]->locked() = true;
 
+        // Take the compose inverse of the locked measurement w.r.t the state transforms
+        const auto compose_inv = steam::se3::ComposeInverseEvaluator::MakeShared(measurement_vars[i], pose_state_vars[i]);
 
-        // Experimental Hard Lateral Constraints
-        /*
-        CLOG(DEBUG, "debug") << "Attempting to add a lateral cost term";
-        const auto inv_state = steam::se3::InverseEvaluator::MakeShared(pose_state_vars[i]);
-        const auto pose_err = steam::vspace::MatrixMultEvaluator<4,4>::MakeShared(inv_state, measurements[i]); // I think this line is the problem, we cannot really convert the se3 evaluators to vspace evaluators
-        const auto left_mul = steam::vspace::MatrixMultEvaluator<1,4>::MakeShared(pose_err, I_2_tran);
-        const auto right_mul = steam::vspace::MatrixMultEvaluator<4,1>::MakeShared(pose_err, I_4);
-        const auto lat_cost_term = steam::WeightedLeastSqCostTerm<6>::MakeShared(right_mul, sharedLatNoiseModel, sharedLossFunc);
-        opt_problem.addCostTerm(lat_cost_term);
-        CLOG(DEBUG, "debug") << "Successfully added Lat cost term";
-        */
+        // Use the ComposeLandmarkEvaluator to right multiply the 4th column of the identity matrix to create a 4x1 homogenous point vector with lat,lon,alt error components
+        const auto error_vec = steam::stereo::ComposeLandmarkEvaluator::MakeShared(compose_inv, I_4_eval);
+
+        // Build lateral barrier terms by querying the current cbit corridor
+        Eigen::Matrix<double, 4, 1> barrier_right;
+        barrier_right <<  0,
+                          barrier_q_right[i],
+                          0,
+                          1;
+        Eigen::Matrix<double, 4, 1> barrier_left;
+        barrier_left <<   0.0,
+                          barrier_q_left[i],
+                          0,
+                          1;
+
+        CLOG(INFO, "mpc_debug.cbit") << "Left Barrier for this meas is: " << barrier_q_left[i];
+        CLOG(INFO, "mpc_debug.cbit") << "Right Barrier for tis meas is: " << barrier_q_right[i];
+        //CLOG(INFO, "mpc_debug.cbit") << "The Initaial Pose is:" << T0;
+        //CLOG(INFO, "mpc_debug.cbit") << "The cbit measurement for this value is: " << measurements_cbit[i].inverse();
+        //CLOG(INFO, "mpc_debug.cbit") << "The vtr measurement for this value is: " << measurements[i].inverse();
+
+        // compute the lateral error using a custom Homogenous point error STEAM evaluator
+        const auto lat_error_right = steam::LateralErrorEvaluatorRight::MakeShared(error_vec, barrier_right); // TODO, rename this evaluator to something else
+        const auto lat_error_left = steam::LateralErrorEvaluatorLeft::MakeShared(error_vec, barrier_left);
+
+        // Previously used Log barriers, however due to instability switch to using inverse squared barriers
+        //const auto lat_barrier_right = steam::vspace::ScalarLogBarrierEvaluator<1>::MakeShared(lat_error_right);
+        //const auto lat_barrier_left = steam::vspace::ScalarLogBarrierEvaluator<1>::MakeShared(lat_error_left);
+
+        // For each side of the barrier, compute a scalar inverse barrier term to penalize being close to the bound
+        const auto lat_barrier_right = steam::vspace::ScalarInverseBarrierEvaluator<1>::MakeShared(lat_error_right);
+        const auto lat_barrier_left = steam::vspace::ScalarInverseBarrierEvaluator<1>::MakeShared(lat_error_left);
+
+        // Generate least square cost terms and add them to the optimization problem
+        const auto lat_cost_term_right = steam::WeightedLeastSqCostTerm<1>::MakeShared(lat_barrier_right, sharedLatNoiseModel, latLossFunc);
+        //opt_problem.addCostTerm(lat_cost_term_right);
+        const auto lat_cost_term_left = steam::WeightedLeastSqCostTerm<1>::MakeShared(lat_barrier_left, sharedLatNoiseModel, latLossFunc);
+        //opt_problem.addCostTerm(lat_cost_term_left);
+
+      
+      }
     }
 
     // Solve the optimization problem with GuassNewton solver
-    using SolverType = steam::GaussNewtonSolver;
-    // Initialize parameters (enable verbose mode)
+    //using SolverType = steam::GaussNewtonSolver; // Old solver, does not have back stepping capability
+    using SolverType = steam::LineSearchGaussNewtonSolver;
+    //using SolverType = steam::DoglegGaussNewtonSolver;
+    // Initialize solver parameters
     SolverType::Params params;
-    params.verbose = false; // Makes the output display for debug when true
+    params.verbose = true; // Makes the output display for debug when true
+    params.relative_cost_change_threshold = 1e-4;
+    params.max_iterations = 100;
+    params.absolute_cost_change_threshold = 1e-4;
+    params.backtrack_multiplier = 0.5; // Line Search Specific Params, will fail to build if using GaussNewtonSolver
+    params.max_backtrack_steps = 400; // Line Search Specific Params, will fail to build if using GaussNewtonSolver
+    //params.ratio_threshold_shrink = 0.1; // Dogleg Specific Params, will fail to build if using GaussNewtonSolver
+    /// Grow trust region if ratio of actual to predicted cost reduction above
+    /// this (range: 0.0-1.0)
+    //params.ratio_threshold_grow = 0.9; // Dogleg Specific Params, will fail to build if using GaussNewtonSolver
+    /// Amount to shrink by (range: <1.0)
+    //params.shrink_coeff = 0.9; // Dogleg Specific Params, will fail to build if using GaussNewtonSolver
+    /// Amount to grow by (range: >1.0)
+    //params.grow_coeff = 1.1; // Dogleg Specific Params, will fail to build if using GaussNewtonSolver
+    /// Maximum number of times to shrink trust region before giving up
+    //params.max_shrink_steps = 200; // Dogleg Specific Params, will fail to build if using GaussNewtonSolver
 
     SolverType solver(opt_problem, params);
+
+    double initial_cost = opt_problem.cost();
+    // Check the cost, disregard the result if it is unreasonable (i.e if its higher then the initial cost)
+    //CLOG(ERROR, "mpc_debug.cbit") << "The Solution Cost is:" << initial_cost;
+
+
+    // Solve the optimization problem
     solver.optimize();
+
+    double final_cost = opt_problem.cost();
+    // Check the cost, disregard the result if it is unreasonable (i.e if its higher then the initial cost)
+    //CLOG(ERROR, "mpc_debug.cbit") << "The Solution Cost is:" << final_cost;
+
+    if (final_cost > initial_cost)
+    {
+      CLOG(ERROR, "mpc_debug.cbit") << "The final cCost was > initial cost, something went wrong. Commanding the vehicle to stop";
+         // First check if any of the values are nan, if so we return a zero velocity and flag the error
+      Eigen::Matrix<double, 2, 1> bad_cost_vel;
+      
+      bad_cost_vel(0) = 0.0;
+      bad_cost_vel(1) = 0.0;
+
+      // if we do detect nans, return the mpc_poses as all being the robots current pose (not moving across the horizon as we should be stopped)
+      std::vector<lgmath::se3::Transformation> mpc_poses;
+      for (int i = 0; i<pose_state_vars.size(); i++)
+      {
+        mpc_poses.push_back(T0);
+      }
+      return {bad_cost_vel, mpc_poses};
+    }
 
     // DEBUG: Display the several of the prediction horizon results
     /*
@@ -270,20 +410,107 @@ struct mpc_result SolveMPC(Eigen::Matrix<double, 2, 1> previous_vel, lgmath::se3
 }
 
 
+/*
+// Measurement generation overhaul, not quite ready to deploy yet
+struct meas_result GenerateTrackingReference(std::shared_ptr<std::vector<Pose>> cbit_path_ptr, std::shared_ptr<CBITPath> global_path_ptr, std::tuple<double, double, double, double, double, double> robot_pose, int K, double DT, double VF, int current_sid)
+{
+  // We start by assuming the robot state is located at the start of the cbit path
+
+  // Save a copy of the current path solution to work on
+  auto cbit_path = *cbit_path_ptr;
+
+  // Iterate through the cbit path until we find the point closest to the true robot pose
+  int closest_pose_ind = 0;
+  double min_dist = INFINITY;
+  for (int i = 0; i <cbit_path.size(); i++)
+  {
+    double dist = sqrt((((cbit_path)[i].x - std::get<0>(robot_pose)) * ((cbit_path)[i].x - std::get<0>(robot_pose))) + (((cbit_path)[i].y - std::get<1>(robot_pose)) * ((cbit_path)[i].y - std::get<1>(robot_pose))));
+    if (dist < min_dist)
+    {
+      closest_pose_ind = closest_pose_ind + 1;
+      min_dist = dist;
+    }
+    else 
+    {
+      // If the minimum dist ever increases, stop. This prevents jumps ahead if the euclidean path doubles back or loops on itself further in the window
+      closest_pose_ind = closest_pose_ind - 1;
+      break;
+    }
+  }
+
+  // Now starting from that closest cbit_path point, calculate chord lengths along the cbit_path_subset up to the horizon length
+  std::vector<double> cbit_p;
+  cbit_chord.reserve(K);
+  double p_dist;
+  cbit_chord.push_back(0.0);
+  for (int j = closest_pose_ind; j < (closest_pose_ind + K); j++)
+  {
+    chord_dist = sqrt((((cbit_path)[j].x - (cbit_path)[j+1].x) * ((cbit_path)[j].x - (cbit_path)[j+1].x)) + (((cbit_path)[j].y - (cbit_path)[j+1].y) * ((cbit_path)[j].y - (cbit_path)[j+1].y)));
+    cbit_chord.push_back(chord_dist + cbit_p.end());
+  }
+
+  //
+
+  // Generate chord length vector along the cbit_path solution
+  double p_dist;
+  std::vector<double> cbit_p;
+  std::vector<Pose> cbit_path_subset;
+  cbit_p.reserve(cbit_path.size());
+  cbit_p.push_back(0.0);
+  for (int i = 0; i < cbit_path.size()-2; i += 1)
+  {
+    chord_dist = sqrt((((cbit_path)[i].x - (cbit_path)[i+1].x) * ((cbit_path)[i].x - (cbit_path)[i+1].x)) + (((cbit_path)[i].y - (cbit_path)[i+1].y) * ((cbit_path)[i].y - (cbit_path)[i+1].y)));
+    cbit_p.push_back(chord_dist);
+    cbit_path_subset.push_pack(cbit_path[i]);
+
+    // Conclude if we reach end of path or exceed the window lookahead range:
+    if (lookahead_dist > 12.0)
+      {
+        break;
+      }
+  }
 
 
 
+  // Generate a subset of points on the euclidean cbit path solution over the current vertex window and find the closest point to the current state
+
+  CLOG(ERROR, "path_planning.cbit_planner") << "Current SID is: " << SID;
+  
+  // Find the corresponding global pose p,q value at the current SID (which should be just behind the actual current state)
+  double current_pq_sid_p = global_path_ptr->p[current_sid];
+  CLOG(ERROR, "path_planning.cbit_planner") << "SID P,Q Pose is: " << " p: " << current_pq_sid_p << " q: " << 0.0;
+
+  std::vector<Node> euclid_subset;
+
+  // The length of the subset is determined by the configuration parameter lookahead distance and the desired discretization
+  // Use the SID values to help predict how far we should look ahead look ahead without falling into false minim on crossing paths.
+  double lookahead_range = cbit_path[0].p + 5.0; //default value from config
+  for (int ID = SID; ID < global_path_ptr->p.size(); ID += 1)
+  {
+    if (cbit_path[0].p < global_path_ptr->p[ID])
+    {
+      lookahead_range = global_path_ptr->p[ID];
+    }
+  }
+
+  for (double i = cbit_path[0].p; i < lookahead_range; i += (1.0 / (20.0)))
+  {
+    euclid_subset.push_back(curve_to_euclid(Node(i,0)));
+    
+  }
+}
+
+*/
 
 
 // helper function for computing the optimization reference poses T_ref based on the current path solution
-struct meas_result GenerateReferenceMeas(std::shared_ptr<std::vector<Pose>> cbit_path_ptr, std::tuple<double, double, double, double, double, double> robot_pose, int K, double DT, double VF)
+// This is specifically for the tracking mpc, but is also used to generate the warm start poses for the corridor mpc
+struct pose_result_tracking GenerateTrackingReference(std::shared_ptr<std::vector<Pose>> cbit_path_ptr, std::tuple<double, double, double, double, double, double> robot_pose, int K, double DT, double VF)
 {
 
 
     // Save a copy of the current path solution to work on
     auto cbit_path = *cbit_path_ptr;
-
-    // Experimental Changes to improve controller instability (completed but not rigourously field tested yet)
 
     // PSEUDO CODE:
     // 1. Find the closest point on the cbit path to the current state of the robot
@@ -301,6 +528,7 @@ struct meas_result GenerateReferenceMeas(std::shared_ptr<std::vector<Pose>> cbit
     double dx;
     double dy;
     double p_correction = 0.0;
+    bool min_flag = true;
 
     std::vector<double> cbit_p;
     cbit_p.reserve(cbit_path.size());
@@ -316,11 +544,20 @@ struct meas_result GenerateReferenceMeas(std::shared_ptr<std::vector<Pose>> cbit
       dx = (cbit_path)[i].x - std::get<0>(robot_pose);
       dy = (cbit_path)[i].y - std::get<1>(robot_pose);
       new_dist = sqrt((dx * dx) + (dy * dy));
-      if (new_dist < min_dist)
+      if ((new_dist < min_dist) && (min_flag == true))
       {
         CLOG(DEBUG, "mpc_debug.cbit") << "Minimum Distance: " << new_dist;
         p_correction = lookahead_dist;
         min_dist = new_dist;
+        //CLOG(ERROR, "mpc_debug.cbit") << "New Minimum Found";
+      }
+      else
+      {
+        //CLOG(ERROR, "mpc_debug.cbit") << "No New Minimum";
+        if (new_dist > min_dist + 0.1)
+        {
+          min_flag = false;
+        }
       }
       
       // Stop once we get about 12m ahead of the robot (magic number for now, but this is just a conservative estimate of any reasonable lookahead window and mpc horizon)
@@ -329,30 +566,33 @@ struct meas_result GenerateReferenceMeas(std::shared_ptr<std::vector<Pose>> cbit
         break;
       }
     }
-    //CLOG(DEBUG, "debug") << "cbit_p is: " << cbit_p;
+    //CLOG(DEBUG, "mpc_debug.cbit") << "cbit_p is: " << cbit_p;
 
     // Determine the p_values we need for our measurement horizon, corrected for the p value of the closest point on the path to the current robot state
     std::vector<double> p_meas_vec;
     std::vector<lgmath::se3::Transformation> measurements;
+    std::vector<double> p_interp_vec;
+    std::vector<double> q_interp_vec;
+
     p_meas_vec.reserve(K);
     for (int i = 0; i < K; i++)
     {
 
       p_meas_vec.push_back((i * DT * VF) + p_correction);
     }
-    //CLOG(DEBUG, "debug") << "p_meas_vec is: " << p_meas_vec;
-
-    // todo: Iterate through the p_measurements and interpolate euclidean measurements from the cbit_path and the corresponding cbit_p values
-    // Note this could be combined in the previous loop too
+    //CLOG(DEBUG, "mpc_debug.cbit") << "p_meas_vec is: " << p_meas_vec;
+    
+    // Iterate through the p_measurements and interpolate euclidean measurements from the cbit_path and the corresponding cbit_p values
+    // Note this could probably just be combined in the previous loop too
     bool point_stabilization = false;
     for (int i = 0; i < p_meas_vec.size(); i++)
     {
       // handle end of path case:
       // if the p meas we would have needed exceeds the final measurement pose, set it equal to our last p value in the path
       // This will cause the intepolation to return the final cbit_path pose for all measurements past this point
-      //CLOG(INFO, "debug") << "The specific p_meas_vec[i] is: " << p_meas_vec[i];
-      //CLOG(INFO, "debug") << "The size of cbit_p is:" << cbit_p.size();
-      //CLOG(INFO, "debug") << "The final cbit_p value is:" << cbit_p[cbit_p.size()-1];
+      //CLOG(INFO, "mpc_debug.cbit") << "The specific p_meas_vec[i] is: " << p_meas_vec[i];
+      //CLOG(INFO, "mpc_debug.cbit") << "The size of cbit_p is:" << cbit_p.size();
+      //CLOG(INFO, "mpc_debug.cbit") << "The final cbit_p value is:" << cbit_p[cbit_p.size()-1];
 
       if (p_meas_vec[i] > cbit_p[cbit_p.size()-1])
       {
@@ -360,22 +600,77 @@ struct meas_result GenerateReferenceMeas(std::shared_ptr<std::vector<Pose>> cbit
         point_stabilization = true; // Enable point stabilization configs
         CLOG(INFO, "mpc.cbit") << "Approaching End of Path, Converting MPC to Point Stabilization Problem";
       }
-      lgmath::se3::Transformation meas = InterpolateMeas(p_meas_vec[i], cbit_p, cbit_path);
+      struct interp_result meas = InterpolateMeas(p_meas_vec[i], cbit_p, cbit_path);
 
       // add to measurement vector
-      measurements.push_back(meas);
+      measurements.push_back(meas.measurement);
+      p_interp_vec.push_back(meas.p_interp);
+      q_interp_vec.push_back(meas.q_interp);
+
     }
 
-    return {measurements, point_stabilization};
+    //CLOG(WARNING, "mpc.cbit") << "The Approximate Robot State P value is (ref meas): " << p_correction;
+    //CLOG(WARNING, "mpc.cbit") << "The p_interp_vec is: " << p_interp_vec;
+    //CLOG(WARNING, "mpc.cbit") << "The q_interp_vec is: " << q_interp_vec;
 
-
-    //End of Experimental Changes
-
+    return {measurements, point_stabilization, p_interp_vec, q_interp_vec};
 }
 
-// function takes in a the cbit path solution with a vector defining hte p axis of the path, and then a desired p_meas
+
+// For generating VT&R teach path measurements used in the corridor mpc (new version which directly uses the interpolated p measurements from the cbit path trajectory tracking)
+struct pose_result_homotopy GenerateHomotopyReference(std::shared_ptr<CBITPath> global_path_ptr, std::shared_ptr<CBITCorridor> corridor_ptr, std::tuple<double, double, double, double, double, double> robot_pose, int K, double DT, double VF, int current_sid, std::vector<double> p_interp_vec)
+{
+    // Set point stabilization, but just note if we use this function in the cbit.cpp file we need to use the Tracking reference pose point stabilization instead
+    bool point_stabilization = false;
+
+    // Initialize vectors storing the barrier values:
+    std::vector<double> barrier_q_left;
+    std::vector<double> barrier_q_right;
+    
+    // load the teach path
+    std::vector<Pose> teach_path = global_path_ptr->disc_path;
+    std::vector<double> teach_path_p = global_path_ptr->p;
+    
+
+    std::vector<lgmath::se3::Transformation> measurements;
+    
+    //CLOG(ERROR, "mpc_debug.cbit") << "The p_interp_vec is: " << p_interp_vec;
+    //CLOG(ERROR, "mpc_debug.cbit") << "The Global Path p is: " << teach_path_p;
+
+    // Iterate through the interpolated p_measurements and make interpolate euclidean measurements from the teach path
+    for (int i = 0; i < p_interp_vec.size(); i++)
+    {
+
+      struct interp_result meas = InterpolateMeas(p_interp_vec[i], teach_path_p, teach_path);
+      //CLOG(WARNING, "corridor_mpc_debug.cbit") << "Adding Measurement: " << meas;
+
+      // add to measurement vector
+      measurements.push_back(meas.measurement);
+
+
+      // Find the corresponding left and right barrier q values to pass to the mpc
+
+      // The corridor_ptr points to the stored barrier values for the entire teach trajectory (0,p_len)
+      // To find the corresponding values, we just need to query the corridor_ptr given the current sid_p + p_meas_vec[i], and return the q values for that bin
+      double p_query = p_interp_vec[i];
+      // this isnt the most efficient way of doing this, but it should be fine, we really only need to run this loop 10-20 times and the size is likely less then 1000 each
+      int p_ind = 0;
+      while (corridor_ptr->p_bins[p_ind] <= p_query)
+      {
+        p_ind++;
+      }
+      barrier_q_left.push_back(corridor_ptr->q_left[p_ind-1]);
+      barrier_q_right.push_back(corridor_ptr->q_right[p_ind-1]);
+      //CLOG(WARNING, "mpc_debug.cbit") << "The left barrier is: " << corridor_ptr->q_left[p_ind-1];
+      //CLOG(WARNING, "mpc_debug.cbit") << "The right barrier is: " << corridor_ptr->q_right[p_ind-1];
+    }
+
+    return {measurements, point_stabilization, barrier_q_left, barrier_q_right};
+}
+
+// function takes in the cbit path solution with a vector defining the p axis of the path, and then a desired p_meas
 // Then tries to output a euclidean pose interpolated for the desired p_meas.
-lgmath::se3::Transformation InterpolateMeas(double p_val, std::vector<double> cbit_p, std::vector<Pose> cbit_path)
+struct interp_result InterpolateMeas(double p_val, std::vector<double> cbit_p, std::vector<Pose> cbit_path)
 {
   // Find the lower bound of the p values
   for (int i = 0; i < cbit_p.size(); i++)
@@ -391,15 +686,38 @@ lgmath::se3::Transformation InterpolateMeas(double p_val, std::vector<double> cb
       Pose pose_lower = cbit_path[i-1];
       Pose pose_upper = cbit_path[i];
 
-      double x_int = pose_lower.x + ((p_val - p_lower) / (p_upper - p_lower)) * (pose_upper .x - pose_lower.x);
-      double y_int = pose_lower.y + ((p_val - p_lower) / (p_upper - p_lower)) * (pose_upper .y - pose_lower.y);
-      double z_int = pose_lower.z + ((p_val - p_lower) / (p_upper - p_lower)) * (pose_upper .z - pose_lower.z);
+    
+      double x_int = pose_lower.x + ((p_val - p_lower) / (p_upper - p_lower)) * (pose_upper.x - pose_lower.x);
+      double y_int = pose_lower.y + ((p_val - p_lower) / (p_upper - p_lower)) * (pose_upper.y - pose_lower.y);
+      double z_int = pose_lower.z + ((p_val - p_lower) / (p_upper - p_lower)) * (pose_upper.z - pose_lower.z);
 
       // For yaw we need to be abit careful about sign and angle wrap around
       // Derive the yaw by creating the vector connecting the pose_upp and pose_lower pts
+      // TODO: There is a problem here for reverse planning, will need to rotate the yaw 180 degrees in that case.
+      // For normal forward planning this is fine though
 
-      double yaw_int = std::atan2((pose_upper.y - pose_lower.y), (pose_upper.x - pose_lower.x));
+      // This interpolation if we do have yaw available (when the input path is the teach path as it is for corridor mpc)
+      double yaw_int;
+      //yaw_int = std::atan2((pose_upper.y - pose_lower.y), (pose_upper.x - pose_lower.x));
+      if ((pose_lower.yaw == 0.0) && (pose_upper.yaw == 0.0))
+      {
+        // Yaw interpolation when we dont have yaw available explicitly (i.e from cbit path euclid conversion)
+        yaw_int = std::atan2((pose_upper.y - pose_lower.y), (pose_upper.x - pose_lower.x));
+        //CLOG(ERROR, "mpc_debug.cbit") << "The Yaw For CBIT PATH is: " << yaw_int;
+      }
+      else
+      {
+        // if this is used on the cbit path, it will just return 0.0 every time (note this will break tracking control if yaw covariance set low)
+        //yaw_int = pose_lower.yaw + ((p_val - p_lower) / (p_upper - p_lower)) * (pose_upper.yaw - pose_lower.yaw);
+        yaw_int = pose_lower.yaw;
+        //CLOG(ERROR, "mpc_debug.cbit") << "The Yaw For TEACH PATH is: " << yaw_int;
+      }
+      
 
+      // we also want to interpolate p and q values based on the original p,q from the cbit_path. We use this afterwards for finding appropriate corridor mpc
+      // reference poses on the teach path
+      double p_int = pose_lower.p + ((p_val - p_lower) / (p_upper - p_lower)) * (pose_upper.p - pose_lower.p);
+      double q_int = pose_lower.q + ((p_val - p_lower) / (p_upper - p_lower)) * (pose_upper.q - pose_lower.q);
 
       // Build the transformation matrix
       Eigen::Matrix4d T_ref;
@@ -412,7 +730,8 @@ lgmath::se3::Transformation InterpolateMeas(double p_val, std::vector<double> cb
       lgmath::se3::Transformation meas = lgmath::se3::Transformation(T_ref);
 
       CLOG(DEBUG, "mpc_debug.cbit") << "The measurement Euclidean state is - x: " << x_int << " y: " << y_int << " z: " << z_int << " yaw: " << yaw_int;
-      return meas;
+      CLOG(DEBUG, "mpc_debug.cbit") << "The measurement P,Q value is - p: " << p_int << " q: " << q_int;
+      return {meas, p_int, q_int};
     }
   }
 }
@@ -425,17 +744,6 @@ Eigen::Matrix<double, 2, 1> SaturateVel(Eigen::Matrix<double, 2, 1> applied_vel,
     double command_ang_z;
     Eigen::Matrix<double, 2, 1> saturated_vel;
 
-    // Moved nan check to the main mpc solver function
-    /*
-    // First check if any of the values are nan, if so we return a zero velocity and flag the error
-    if (std::isnan(applied_vel(0)) || std::isnan(applied_vel(1)))
-    {
-      CLOG(ERROR, "mpc.cbit") << "NAN values detected, mpc optimization failed. Returning zero velocities";
-      saturated_vel(0) = 0.0;
-      saturated_vel(1) = 0.0;
-      return saturated_vel;
-    }
-    */
 
     if (abs(applied_vel(0)) >= v_lim)
     {
@@ -448,11 +756,7 @@ Eigen::Matrix<double, 2, 1> SaturateVel(Eigen::Matrix<double, 2, 1> applied_vel,
         command_lin_x = -1.0* v_lim;
       }
     }
-    // Removed for bi-directional control purposes
-    //else if (applied_vel(0)  <= 0.0)
-    //{
-    //  command_lin_x = 0.0;
-    //}
+
     else
     {
       command_lin_x = applied_vel(0) ;
@@ -469,50 +773,12 @@ Eigen::Matrix<double, 2, 1> SaturateVel(Eigen::Matrix<double, 2, 1> applied_vel,
         command_ang_z = -1.0 * w_lim;
       }
     }
-    //else if (applied_vel(1) <= -1*w_lim)
-    //{
-    //  command_ang_z = -1*w_lim;
-    //}
+
     else
     {
       command_ang_z = applied_vel(1) ;
     }
 
-    // Changes for Bi-directional path traversal, we no longer want to saturate commands less than 0.0
-
     saturated_vel << command_lin_x, command_ang_z;
     return saturated_vel;
 }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
